Arduino Clone
This board is optimized for the the othermill so it may do with a 1/32
bit in a relatively fast time. Bare minimum Materials ATmega328p, 16MHZ
ceramic resonator 0.1uf capacitors , Button(Short two wires if
desperate), 1K resistor and female or male headers. If you want to use
the voltage regulator you will need the 220uf capacitors.